Fix to use the right client to download the image
